Added Layouts
I had to create a layout activity because if we kept it on activity_create_theme, there would be no room to select a theme and view it. So the app runs activity_layout first to test the fragment inflation. And it works. I just have to add the transition from the layout btn on activity_create_theme to activity_layout and will will be able to access it.